Move js files into src folder
