+++ package.json
@@ -7,7 +7,7 @@
     "ios": "react-native run-ios",
     "start": "react-native start",
     "test": "jest",
-    "lint": "eslint ./components ./App.js ./types.js",
+    "lint": "eslint ./src ./App.js",
     "flow": "flow",
     "flow-stop": "flow stop"
   },
